Create Processor encryption
Content encryption helps to protect the confidentiality of content that you add to a storage area in case the content is accessed outside of FileNet® P8. This encryption pertains only to the storage of content in the storage area: when Content Platform Engine retrieves and passes content to a client in response to a client request, the content is automatically decrypted.

Content Platform Engine encrypts and decrypts content using AES in Counter mode, a Federal Information Processing Standard (FIPS) 140-compliant algorithm, with a 128-bit key or a 256-bit key. A new key is generated whenever you enable encryption for the storage area. For example, when you first enable encryption, one encryption key exists, and that key is used to encrypt new content. If you re-enable encryption, two encryption keys now exist, and the most recent key is used to encrypt new content. If you re-enable encryption again, three encryption keys now exist, and so on. The storage area encryption keys are stored in a secure form in the object store database.